UmlClassFigure is actually *doing* something interesting now! (as a demo, try inputting whatever you'd like for class names on class figures. It should reset to something valid no matter what *and* find the access modifier character (as long as its at index 0...)) I've added listeners that will check for any change to the figure and remap the data into model objects as needed. I'm using some regular expressions and a lot of string examination to make sure we're always pumping valid data into the models and maintaining a consistent state between the ui and model.
However, I've run into a snag with the AttributesCompartment not updating
as I'd expect; the listener doesn't seem to be getting fresh data, and
so it's not actually checking edits made by users. The class name text
figure is working though, so I'm not sure why there's this weirdness here.

Signed-off-by: hendleyr <hendleyr@gmail.com>