Implemented drawing associations and removed several extraneous line connection classes. Fixed several major bugs with enums and strings when saving/opening files; the solution is not terribly graceful but it works.
EXISTING ISSUES
-The stupid aggregation decoration is showing up skewed. I spent a while
tinkering at it but my math skills aren't what they used to be.
-I will need to refactor UmlAssociationModel to store additional data: a
reference to the UmlClassModel on which the AssociationModel exists. This
will greatly simplify user input validation on AssociationFigures. For the
moment, it is possible to create inheritance cycles in a number of ways;
this will be fixed when I complete the refactoring and implement controls
in the actions.

Signed-off-by: hendleyr <hendleyr@gmail.com>