Update src/ui/UmlClass/UmlClassFigure.java
Logic for parsing user input for methods should be good, but there seem to be some bugs elsewhere; I get a Source Not Found message in EventDispatchThread whenever adding a new class figure to the diagram. In addition, updating any of the text fields and moving the class figures does not seem to work. I don't think this should have been caused by the MethodAdapter code, so we need to find out where this is coming from.